diplom/requirements.txt
# FastAPI framework and server
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data validation
pydantic==2.4.2

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1